update from reviews
