+++ src/Rule/NonStaticAssertions.php
@@ -14,19 +14,23 @@ declare(strict_types=1);
 namespace App\Rule;
 
 use App\Attribute\Rule\Description;
-use App\Attribute\Rule\InvalidExample;
 use App\Attribute\Rule\ValidExample;
+use App\Attribute\Rule\InvalidExample;
+use App\Rst\RstParser;
+use App\Traits\DirectiveTrait;
 use App\Value\Lines;
-use App\Value\NullViolation;
 use App\Value\RuleGroup;
 use App\Value\Violation;
+use App\Value\NullViolation;
 use App\Value\ViolationInterface;
 
 #[Description('Use `$this->assert*` over static calls.')]
 #[InvalidExample('self::assertTrue($foo);')]
 #[ValidExample('$this->assertTrue($foo);')]
-class UseNonStaticAssertions extends AbstractRule implements LineContentRule
+final class NonStaticAssertions extends AbstractRule implements LineContentRule
 {
+    use DirectiveTrait;
+
     public static function getGroups(): array
     {
         return [
@@ -39,7 +43,18 @@ class UseNonStaticAssertions extends AbstractRule implements LineContentRule
         $lines->seek($number);
         $line = $lines->current();
 
-        if ($line->raw()->match('/self::assert*/') || $line->raw()->match('/static::assert*/')) {
+        //if (!RstParser::isPhpDirective($line)) {
+        //    return NullViolation::create();
+        //}
+
+        //$lines->next();
+        //++$number;
+
+        //echo($lines->current()->raw());
+
+        if ($this->inPhpCodeBlock($lines, $number)
+        && ($lines->current()->raw()->match('/self::assert*/')
+        || $lines->current()->raw()->match('/static::assert*/'))) {
             return Violation::from(
                 'Please use `$this->assert` over static call',
                 $filename,
